refactor(search): cache MiniSearch index
