+++ package.json
@@ -35,6 +35,7 @@
   },
   "dependencies": {
     "@epic-web/invariant": "1.0.0",
+    "@epic-web/remember": "1.1.0",
     "@fontsource/source-sans-3": "5.2.9",
     "@react-router/node": "7.9.5",
     "@react-router/serve": "7.9.5",
